refactor(oauth): separate channels into their own file
